// web-ui/src/app/shared/enum/cms-sytem-enum.ts
export enum CmsSystemEnum {
  Index = 1,
  BookUniqueHome = 2,
  BookHotel = 3,
  BookFlight = 4,
  BookTour = 5,
  BookTaxi = 6,
  BookExperience = 7
}

export enum divCmsId {
  div1 = 13,
  div2 = 14,
  div3 = 15,
  div4 = 16,
  div5 = 17,

  div2Title = 52,
  div3Title = 53,
  div4Title = 54,
  div5Title = 55,
}


export enum AttachmentType {
  Image = 158,
  Video = 159,
  Text = 198,
  Review = 160
}